backend/labroom_service/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
gunicorn==20.1.0
asyncpg==0.28.0
pydantic==2.1.1
pydantic-settings==2.0.1
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
email-validator==2.0.0
pyjwt==2.7.0
aiofiles==23.1.0
httpx==0.24.1
pytest==7.3.1
pytest-asyncio==0.21.0
aiohttp==3.8.5 
reportlab==3.6.12
pillow==9.4.0
websockets==10.4